Users can now join groups on profile page
Made some changes to routes, particularily there was a couple of problems with /groups/addmember. Firstly, for some reason despite the default value for private_type, it wasn't being set so a search for private_type: false gave nothing. Secondly the query appeared to query groupmembers for instances that already had that user as a member, and then update it to have that user be a member of the group, which doesn't make much sense. I also updated users/user/groups to also give description.